Add settings for new MacBook Pro
When docked it is using a standard 2560-width monitor, but none of
those sizes are good for the laptop's default 1512-width screen.
Start with a left/right split that works for Terminal/text editor,
and note that the bottom row of the keyboard should be reserved
for future per-machine sizes.