Add WiiM controls to the Stream Deck
Although when the music is being controlled by an app running on the
Mac the standard macOS play/pause/skip buttons work, the volume change
is wrong. Having dedicated WiiM buttons for playback and volume works,
as well as when the music is coming from my phone not my desktop.