Add AntiRSI controls to the Stream Deck
`antirsi` is a script I use to enforce that the app AntiRSI
(https://antirsi.onnlucky.com) is running because otherwise I will
absent-mindedly quit it to stop it bugging me. I want it to keep
returning. It also has options to turn it off (only for a few hours)
so I can play a game without being bothered.

This provides a status and control for that script in my deck.