Add a scene cycler to Stream Deck
A simple HomeKit button sets a scene when pressed.

A list step button sets the next button image and runs a callback when
pressed, cycling back to the beginning at the end of the list.

I'm using the list step to set a scene and turn the key light on/off.